+++ packages/kb-client/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @kb/client
 
+## 1.3.8
+
+### Patch Changes
+
+- Updated dependencies
+  - @kb/core@1.5.2
+
 ## 1.3.7
 
 ### Patch Changes